Don't merge and symlink macOS app renderer/game libraries
The engine only references the separate filenames, there is no advantage
to including merged libraries without the arch in the filename.

This also reduces file size when the app bundle is zipped without
specifying to keep symlinks. The script merged multiple arches to one
file and then creating the zip caused the file to be duplicated for each
arch and the archless filename.